feat: Return info with `/bot/builds` & `/bot/runs`

+++ backend/df_designer/app/schemas/preset.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
 
 class Preset(BaseModel):
-    wait_time: int
+    wait_time: float
     end_status: str
